+++ include/util/compression/int/fastpfor/variablebyte.h
@@ -14,15 +14,15 @@ class VariableByte: public IntegerCODEC
 {
 public:
     template<uint32_t i>
-    uint32_t extract7bits(const uint32_t val)
+    uint8_t extract7bits(const uint32_t val)
     {
-        return (val >> (7 * i)) & ((1U << 7) - 1);
+        return static_cast<uint8_t>((val >> (7 * i)) & ((1U << 7) - 1));
     }
 
     template<uint32_t i>
-    uint32_t extract7bitsmaskless(const uint32_t val)
+    uint8_t extract7bitsmaskless(const uint32_t val)
     {
-        return (val >> (7 * i));
+        return static_cast<uint8_t>((val >> (7 * i)));
     }
 
     void encodeArray(const uint32_t *in, const size_t length, uint32_t *out,
@@ -39,7 +39,7 @@ public:
              */
             if (val < (1U << 7))
             {
-                *bout = val | (1U << 7);
+                *bout = static_cast<uint8_t>(val | (1U << 7));
                 ++bout;
             }
             else if (val < (1U << 14))
